+++ src/components/AddData/AddData.stories.js
@@ -3,25 +3,19 @@ import AddData from './AddData';
 import Map from '../Map/Map';
 import ElementsProvider from '../_common/ElementsProvider';
 import { storiesOf } from '@storybook/react';
-import { withKnobs } from '@storybook/addon-knobs';
+import { withKnobs, radios } from '@storybook/addon-knobs';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import mapOptions from '../../util/mockMapOptions';
 
 storiesOf('AddData', module)
   .addDecorator(withKnobs)
-  .add('In Panel', () => {
+  .add('Default', () => {
     return (
       <ElementsProvider>
-        <Map mapOptions={mapOptions} />
-        <AddData panel />
-      </ElementsProvider>
-    );
-  })
-  .add('Off Panel', () => {
-    return (
-      <ElementsProvider>
-        <AddData />
-        <Map mapOptions={mapOptions} />
+        <Map mapOptions={mapOptions}></Map>
+        <AddData
+          baseType={radios('baseType', ['none', 'panel', 'button'], 'panel')}
+        />
       </ElementsProvider>
     );
   });