+++ src/components/Bookmarks/Bookmarks.stories.js
@@ -3,10 +3,14 @@ import Bookmarks from './Bookmarks';
 import Map from '../Map/Map';
 import ElementsProvider from '../_common/ElementsProvider';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, boolean } from '@storybook/addon-knobs';
+import { withKnobs, radios } from '@storybook/addon-knobs';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import mapOptions from '../../util/mockMapOptions';
-import { loadBookmarks, deleteBookmark, saveBookmark } from './util/bookmarkActions';
+import {
+  loadBookmarks,
+  deleteBookmark,
+  saveBookmark
+} from './util/bookmarkActions';
 
 const goToBookmark = (map, bookmark) => {
   map.flyTo({
@@ -15,45 +19,37 @@ const goToBookmark = (map, bookmark) => {
     bearing: 0,
     speed: 0.5,
     curve: 1.7,
-    easing: function(t) {
+    easing: function (t) {
       return t;
     },
     essential: true
   });
   return true;
-}
-
+};
 
 storiesOf('Bookmarks', module)
   .addDecorator(withKnobs)
-  .add('Default (In Panel)', () => {
+  .add('Default', () => {
     return (
       <ElementsProvider>
         <Map mapOptions={mapOptions} />
-        <Bookmarks panel={true} />
-      </ElementsProvider>
-    );
-  })
-  .add('Out of Panel', () => {
-    return (
-      <ElementsProvider>
-        <Bookmarks panel={false} />
-        <Map mapOptions={mapOptions} />
+        <Bookmarks
+          baseType={radios('baseType', ['none', 'panel', 'button'], 'panel')}
+        />
       </ElementsProvider>
     );
   })
   .add('Custom actions', () => {
     return (
       <ElementsProvider>
-        <Bookmarks 
-          panel={true}
-          goToBookmark={goToBookmark} 
-          loadBookmarks={loadBookmarks} 
-          deleteBookmark={deleteBookmark} 
-          saveBookmark={saveBookmark} 
+        <Bookmarks
+          baseType={radios('baseType', ['none', 'panel', 'button'], 'panel')}
+          goToBookmark={goToBookmark}
+          loadBookmarks={loadBookmarks}
+          deleteBookmark={deleteBookmark}
+          saveBookmark={saveBookmark}
         />
         <Map mapOptions={mapOptions} />
       </ElementsProvider>
-    )
-  })
-  ;
+    );
+  });