+++ src/components/Draw/Draw.stories.js
@@ -3,7 +3,7 @@ import Draw from './Draw';
 import Map from '../Map/Map';
 import ElementsProvider from '../_common/ElementsProvider';
 import { storiesOf } from '@storybook/react';
-import { withKnobs } from '@storybook/addon-knobs';
+import { withKnobs, radios } from '@storybook/addon-knobs';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import mapOptions from '../../util/mockMapOptions';
 
@@ -18,7 +18,9 @@ storiesOf('Draw', module)
     return (
       <ElementsProvider>
         <Map mapOptions={customMapOptions} />
-        <Draw right="1rem" />
+        <Draw
+          baseType={radios('baseType', ['none', 'panel', 'button'], 'none')}
+        />
       </ElementsProvider>
     );
   });