+++ src/components/MapInfo/MapInfo.stories.js
@@ -3,7 +3,7 @@ import MapInfo from './MapInfo';
 import Map from '../Map/Map';
 import ElementsProvider from '../_common/ElementsProvider';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, boolean, number } from '@storybook/addon-knobs';
+import { withKnobs, radios } from '@storybook/addon-knobs';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import mapOptions from '../../util/mockMapOptions';
 
@@ -13,7 +13,9 @@ storiesOf('MapInfo', module)
     return (
       <ElementsProvider>
         <Map mapOptions={mapOptions}>
-          <MapInfo></MapInfo>
+          <MapInfo
+            baseType={radios('baseType', ['none', 'panel', 'button'], 'none')}
+          />
         </Map>
       </ElementsProvider>
     );