+++ src/components/Zoom/Zoom.js
@@ -40,12 +40,15 @@ const Zoom = ({
 
   let buttonStyle = {
     borderRadius: 'default',
+    width: '32px',
+    height: '32px',
     minWidth: '25px',
     minHeight: '25px',
+    boxShadow: '0 0 4px rgba(0, 0, 0, .6)',
     padding: 0,
-    margin: '1px',
+    my: '2px',
     '&:hover': {
-      bg: 'accent'
+      bg: 'secondary'
     }
   };
 
@@ -56,10 +59,22 @@ const Zoom = ({
     };
   }
 
+  if (controlType === 'slider') {
+    buttonStyle = {
+      ...buttonStyle,
+      mx: 0
+    };
+  }
+
   let flexDirection = 'column';
 
   if (horizontal) {
     flexDirection = 'row';
+    buttonStyle = {
+      ...buttonStyle,
+      marginRight: '4px',
+      my: 0
+    };
   }
 
   let zoomLevelIndicator = null;
@@ -154,7 +169,7 @@ const Zoom = ({
             max={22}
             sx={{
               width: '120px',
-              my: 5,
+              my: '58px',
               transform: 'rotate(270deg)'
             }}
             value={zoom}
@@ -177,9 +192,12 @@ const Zoom = ({
   return (
     <BaseComponent
       {...rest}
+      baseType="none"
       sx={{
         display: 'flex',
-        flexDirection
+        flexDirection,
+        m: 1,
+        ...sx
       }}
       className="cl-zoom-controls"
     >