+++ src/components/Zoom/Zoom.stories.js
@@ -1,11 +1,14 @@
 import React from 'react';
 import Zoom from './Zoom';
+import Bookmarks from '../Bookmarks/Bookmarks';
 import Map from '../Map/Map';
+import Home from '../Home/Home';
 import ElementsProvider from '../_common/ElementsProvider';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, boolean, number } from '@storybook/addon-knobs';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import mapOptions from '../../util/mockMapOptions';
+import { Flex } from 'theme-ui';
 
 storiesOf('Zoom', module)
   .addDecorator(withKnobs)
@@ -14,7 +17,6 @@ storiesOf('Zoom', module)
       <ElementsProvider>
         <Map mapOptions={mapOptions} />
         <Zoom
-          sx={{ bg: 'secondary' }}
           circular={boolean('circular', false)}
           horizontal={boolean('horizontal', false)}
           showZoomLevel={boolean('showZoomLevel', false)}
@@ -27,11 +29,32 @@ storiesOf('Zoom', module)
       <ElementsProvider>
         <Map mapOptions={mapOptions} />
         <Zoom
-          sx={{ bg: 'secondary' }}
           circular={boolean('circular', false)}
           horizontal={boolean('horizontal', false)}
           controlType="slider"
         />
+      </ElementsProvider>
+    );
+  })
+  .add('Misc', () => {
+    return (
+      <ElementsProvider>
+        <Map mapOptions={mapOptions}>
+          <Flex
+            sx={{ flexDirection: 'column', justifyContent: 'space-evenly' }}
+          >
+            <Zoom
+              sx={{ position: 'relative' }}
+              circular={boolean('circular', false)}
+              horizontal={boolean('horizontal', false)}
+            />
+            <Home
+              sx={{ position: 'relative' }}
+              initBounds={mapOptions.bounds}
+            />
+            <Bookmarks baseType="button" sx={{ position: 'relative' }} />
+          </Flex>
+        </Map>
       </ElementsProvider>
     );
   });