+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Upcoming
+
+**BREAKING CHANGES:**
+
+- `<BaseComponent/>` API has been updated.
+- Removed the following props: `panel`, `top`, `left`, `bottom`
+- Added the following props:
+  - `baseType` - choices are `none` (default), `panel`, `button`
+  - `sx` - set of styles that are passed to the base container (this existed but is now passed consistently)
+  - `buttonOptions` - set of options passed to button-style components
+- New <ButtonComponent> (`baseType = 'button`) provides and out-of-the-box UI option for button triggers
+
 # 1.3.7
 
 - Add title to Home button